extracting logical function to a new workflow file
