+++ src/components/display/fewlines/ValidateIdentityForm/ValidateIdentityForm.tsx
@@ -10,39 +10,12 @@ import { Form } from "../Form/Form";
 import { Input } from "../Input/Input";
 import { NeutralLink } from "../NeutralLink/NeutralLink";
 import { IdentityTypes } from "@lib/@types/Identity";
-import { HttpVerbs } from "@src/@types/core/HttpVerbs";
 import {
   InvalidValidationCode,
   TemporaryIdentityExpired,
 } from "@src/clientErrors";
 import { displayAlertBar } from "@src/utils/displayAlertBar";
-import { fetchJson } from "@src/utils/fetchJson";
-
-export async function validateIdentity(
-  validationCode: string,
-  eventId: string,
-): Promise<string> {
-  return fetchJson("/api/auth-connect/verify-validation-code", HttpVerbs.POST, {
-    validationCode,
-    eventId,
-  }).then(async (response) => {
-    if (response.status >= 400) {
-      const { error } = await response.json();
-
-      if (error === "INVALID") {
-        throw new InvalidValidationCode("Invalid validation code");
-      }
-
-      if (error === "Temporary Identity Expired") {
-        throw new TemporaryIdentityExpired();
-      }
-    }
-
-    const path = new URL(response.url).pathname;
-
-    return path;
-  });
-}
+import { validateIdentity } from "@src/workflows/validateIdentity";
 
 export const ValidateIdentityForm: React.FC<{
   type: IdentityTypes;
